separate out tokenRanking into a tokenSelector class -> single responsibility principle
